Make Fl_Gl_Window::pixels_per_unit() return a float (rather than int) value.
This is compatible with future scaling factors applicable to the GUI.

git-svn-id: file:///fltk/svn/fltk/branches/branch-1.3@11787 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ src/Fl_Gl_Window.cxx
@@ -96,7 +96,7 @@ void Fl_Gl_Window::show() {
 
 #if defined(__APPLE__)
 
-int Fl_Gl_Window::pixels_per_unit()
+float Fl_Gl_Window::pixels_per_unit()
 {
   return (fl_mac_os_version >= 100700 && Fl::use_high_res_GL() && Fl_X::i(this) && Fl_X::i(this)->mapped_to_retina()) ? 2 : 1;
 }
